requirements.txt
absl-py==0.11.0
appdirs==1.4.4
cachetools==4.2.1
certifi==2020.12.5
cfgv==3.1.0
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
google-auth==1.27.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
identify==1.5.13
idna==2.10
importlib-metadata==3.4.0
Markdown==3.3.3
nodeenv==1.5.0
numpy==1.20.1
oauthlib==3.1.0
Pillow==8.1.0
pre-commit==2.10.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.2
Werkzeug==1.0.1
yacs==0.1.8
zipp==3.4.0
